+++ projects/swimlane/ngx-charts/src/lib/pie-chart/pie-grid-series.component.ts
@@ -9,7 +9,8 @@ import {
   ChangeDetectionStrategy
 } from '@angular/core';
 import { pie } from 'd3-shape';
-import { PieArc, PieGridData } from '../models/chart-data.model';
+import { PieGridData } from '../models/chart-data.model';
+import { PieArc } from './pie-grid.component';
 
 @Component({
   selector: 'g[ngx-charts-pie-grid-series]',
